successfully added next-auth
