+++ lib/db.ts
@@ -1,7 +1,7 @@
 import { Pool } from "pg";
 
 let db;
-const connectionString = 'postgres://pxcddmbe:soXd8-HF0Fz9__fNMoqeFXTq3upUEnWN@heffalump.db.elephantsql.com/pxcddmbe';
+const connectionString = process.env.CONNECTION_STRING;
 
 if (!db) {
   db = new Pool({
