+++ pages/_app.tsx
@@ -1,0 +1,14 @@
+import './globals.css'
+import { SessionProvider } from "next-auth/react";
+import type { AppProps } from 'next/app'
+
+export default function App({
+	Component,
+	pageProps: { session, ...pageProps },
+}) {
+	return (
+		<SessionProvider session={session}>
+			<Component {...pageProps} />
+		</SessionProvider>
+	);
+}
