--- pages/api/goals.ts
@@ -1,2 +1,0 @@
-import { NextApiRequest, NextApiResponse } from 'next'
-import db from '../../lib/db'
